Renaming registry.hpp to registry_base.hpp in order to avoid naming conflicts.
