+++ src/util/registry/registry_impl.hpp
@@ -2,7 +2,7 @@
 
 #pragma once
 
-#include "registry.hpp"
+#include "registry_base.hpp"
 #include "cavl.hpp"  // Copy from https://github.com/pavel-kirienko/cavl/blob/main/c%2B%2B/cavl.hpp
 #include <memory>
